Start from a working example
Gustavo created a working example of how to use libcurl as the basis of
a websocket.  With his blessing this library's goal is to take his code
and turn it into a library that is easy to include and use in other
programs.

Below is the link where the code was effectively forked.

https://github.com/barbieri/barbieri-playground/commit/ddf033aa338c07ebbfba5ed6ce08d586e1213993

The author is set to Gustavo as he wrote this code originally.