A few changes to the API
The few small-ish changes to the API had a larger set of changes to the
rest of the tree.

 * Removed the debug function pointer argument from the API
 * Removed the get_random function pointer argument from the API

Both of these are better served at compile time and just bloated the
API.

 * Removed the WS close codes.

Curl doesn't have macros for things like HTTP codes, so I think we
should be silent here as well.

 * Moved the sha1 code into it's own file.
 * Moved random into it's own file.

Both of these moves are towards compile time changes later so that
someone can replace these calls simply.

 * Added a cws_close_urgent()

This helps address the need for closing a session without delay.