Get all autobahn tests to pass.
Still work in progress, but this gets all the autobahn tests to pass.
To do this, a bunch of UTF8 work had to be done as well as several
refactors to handle the streaming in a useful way.  There is still a
fair bit of work, but this is relatively working code.