Improve tests and coverage
Fixed the test_receive code and made it surprisingly effective with a
very small number of tests that only use the callbacks.  This codes
passes all the autobahn tests and fixes a few defects realted to stream
processing.  This is still work in progress as the receive logic seems
too complex as of now.