Add tls_version and fix follow_redirects
In order to debug why the redirection code wasn't working as expected I
needed to change the TLS version to 1.2 so I could debug using
wireshark.  In the process there were several other improvements to
loggin made as well.

  * TLS upgrade/downgrade is working.
  * HTTP redirects are being followed properly & also being ignored properly.

One area not tested is the downgrade to an older version of libcurl.