Refactor the close code per websocket spec
This change enables the code to ensure a close frame is sent and
received, then after both, the connection is closed.  Due to the likely
ongoing debugging needed a verbose_close() call was made to simplfy the
output of the connection state.

Overall this works with a real test server and also works when the cord
is pulled from the connection (a 56 after a timeout occurs).